Improving the quality of treeComparison Functions implementation.
The treeComparison test class method is now void due to the fact that it now directly uses the Qt QCOMPARE macros. In line with this, boolean flags have also been removed from inside the function implementation. The implementations of the test methods have also undergone changes in accordance with changes in the work of the comparison function.